Rename GeoMapMenuListElement
